% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bow_ci.R
\name{bow_ci}
\alias{bow_ci}
\title{Compute bootstrap CIs for the average descriptor occurrence metric of the bow_analysis function}
\usage{
bow_ci(
  bow_analysis_output,
  bootstraps = 1000,
  alpha = 0.95,
  window = 10,
  per_occurrence = TRUE,
  bootstrap_terms = TRUE
)
}
\arguments{
\item{bow_analysis_output}{result list outputted by the bow_analysis function}

\item{bootstraps}{number of bootstrap samples to draw (default = 1000)}

\item{alpha}{alpha value to compute CIs with}

\item{window}{window size used in the bow_analysis function (default = 10)}

\item{per_occurrence}{per occurrence value used in the bow_analysis function (default = TRUE)}

\item{bootstrap_terms}{if TRUE the bootstrap function samples the set of all terms within a window of all phenomenon occurrences. If FALSE the boostrap function samples from a vector of descriptor counts for each phenomenon occurrence.}
}
\value{
list of analzed texts. Each text list contains a data frame of observed means and CI bounds for each descriptor and phenomenon pair.
}
\description{
Compute bootstrap CIs for the average descriptor occurrence metric of the bow_analysis function
}
