% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_plot.R
\name{ci_plot}
\alias{ci_plot}
\title{ci_plot}
\usage{
ci_plot(plot_data, scale = "shape")
}
\arguments{
\item{plot_data}{output of create_plot_data function}

\item{scale}{character vector specifying whether phenomena should be plotted with different colours or shapes. options are "shape" or "colour".}
}
\value{
a ggplot object
}
\description{
plot results of bow_analysis with bow_ci confidence intervals
}
