% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_analysis.R
\name{dfm_analysis}
\alias{dfm_analysis}
\title{Create a document frequency matrix (sorted in descending order) of terms that occur within some window around a given set of words (phenomenon)}
\usage{
dfm_analysis(
  corpus,
  phenomenon,
  window = 10,
  n_terms = 10,
  filter_dictionary = NULL,
  tf_idf = FALSE,
  filter_ps = FALSE,
  ps = NULL,
  own_regex = FALSE
)
}
\arguments{
\item{corpus}{the text or texts to be analyzed as a list of character vectors}

\item{phenomenon}{a list of character vectors (or list of regular expressions if own_regex == TRUE) with terms around which words will be counted for the dfm}

\item{window}{number of words left and right of a phenomenon term to be considered for the dfm}

\item{n_terms}{number of terms displayed in dfm}

\item{filter_dictionary}{a character vector (or regular expression if own_regex == TRUE) of words to select from the dfm}

\item{tf_idf}{if TRUE function computes tf-idf metric instead of raw counts}

\item{filter_ps}{if TRUE enables filtering of results by part of speech (i.e only adjectives and adverbs)}

\item{ps}{character vector of parts of speech to filter. see selection with unique(tidytext::parts_of_speech[,"pos"])}

\item{own_regex}{when TRUE allows you to add custom regular expressions for phenomenon and filter_dictionary. when FALSE rbow will construct regular expression from the character vectors you supplied. defaults to FALSE}
}
\value{
list of dfms (one dfm per text in corpus)
}
\description{
Create a document frequency matrix (sorted in descending order) of terms that occur within some window around a given set of words (phenomenon)
}
