% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_construct.R
\name{grep_construct}
\alias{grep_construct}
\title{utility function that creates regular expression for grep calls in bow and dfm analysis}
\usage{
grep_construct(text_input, collapse = FALSE)
}
\arguments{
\item{text_input}{a character vector or list of character vectors to be turned into regular expression}

\item{collapse}{when TRUE function collapses text_input into a single regex string. when FALSE function creates a vector or regex strings. defaults to FALSE}
}
\value{
a list of regular expressions if input was a list or a regular expression character vector if input was vector
}
\description{
utility function that creates regular expression for grep calls in bow and dfm analysis
}
