% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{tf_idf}
\alias{tf_idf}
\title{find the most relevant document for a queery by computing tf-idf for a set of terms within a corpus}
\usage{
tf_idf(corpus, terms, own_regex = FALSE)
}
\arguments{
\item{corpus}{the text or texts to be analyzed as a list of character vectors}

\item{terms}{a character vector (or a regular expression if own_regex == TRUE) of terms to determine document relevance with}

\item{own_regex}{when TRUE allows you to add custom regular expressions for terms argument. when FALSE rbow will construct regular expression from the character vector you supplied. defaults to FALSE}
}
\value{
a data frame with document number and mean tf-idf
}
\description{
find the most relevant document for a queery by computing tf-idf for a set of terms within a corpus
}
