% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{clean_text}
\usage{
clean_text(texts, lexicon = "SMART")
}
\arguments{
\item{texts}{list of character vectors}

\item{lexicon}{stop word lexicon to use}
}
\value{
list of cleaned character vectors
}
\description{
clean_text
}
