% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{clean text data (remove punctuation, special characters, numbers, english stopwords)}
\usage{
clean_text(texts, rm_stopwords = FALSE, stopwords_language = NULL)
}
\arguments{
\item{texts}{list of character vectors}

\item{rm_stopwords}{if TRUE clean_text removes stopwords (default FALSE)}

\item{stopwords_language}{stopword dictionary language ISO 639-1 code as character vector}
}
\value{
list of cleaned character vectors
}
\description{
clean text data (remove punctuation, special characters, numbers, english stopwords)
}
