% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_data.R
\name{create_plot_data}
\alias{create_plot_data}
\title{re-format output of bow_ci function to plottable format}
\usage{
create_plot_data(bstrap_output)
}
\arguments{
\item{bstrap_output}{output of bow_ci function}
}
\value{
data frame of plot data
}
\description{
re-format output of bow_ci function to plottable format
}
