% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_texts.R
\name{stem_texts}
\alias{stem_texts}
\title{apply SnowballC stemming function to texts}
\usage{
stem_texts(texts, language = "english")
}
\arguments{
\item{texts}{list of character vectors}

\item{language}{a language supported by SnowballC defaults to english}
}
\value{
list of stemmed character vectors
}
\description{
apply SnowballC stemming function to texts
}
