% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_window.R
\name{text_window}
\alias{text_window}
\title{Utility function that returns the subset of elements in a vector between a start and end position}
\usage{
text_window(text, start, end, position, out_of_window = FALSE)
}
\arguments{
\item{text}{input vector}

\item{start}{start position}

\item{end}{end position}

\item{position}{vector of word positions around which a window is constructed}

\item{out_of_window}{if TRUE function returns set of words within and outside of window}
}
\description{
Utility function that returns the subset of elements in a vector between a start and end position
}
